+++ test/test_issue1.py
@@ -23,4 +23,5 @@ def test_tinker(app):
         app.env.bibtex_cache.cited,
         {u'2012/07/24/hello_world_': set([u"2011:BabikerIPv6"])})
     nose.tools.assert_equal(
-        app.env.bibtex_citation_label, {u"2011:BabikerIPv6": "1"})
+        app.env.bibtex_cache.bibliographies['bibtex-bibliography-0'].labels,
+        {u"2011:BabikerIPv6": "1"})