+++ CHANGELOG.rst
@@ -14,8 +14,8 @@
 
 * Use pybtex labels (fixes issue #6 and issue #7).
 
-* Cache tracked citation keys (fixes issues with citations in repeated
-  Sphinx runs).
+* Cache tracked citation keys and labels (fixes issues with citations
+  in repeated Sphinx runs).
 
 0.2.3 (30 July 2012)
 --------------------